skel: "make debug" does normal make too
There are a couple reasons for doing this.  First, it's annoyingly
easy to forget to enter the debug URL, and you end up loading an
old version of the normal code.  This cures that (but not the
opposite case, which also happens).  Probably more importantly, the
line numbers on compiler errors for the debug version are currently
not always right.  It's probable that could be fixed with some
tweaks to the numberizer, but for now, this gets around it.  It
compiles in normal mode first; if there are compile errors, they
should trigger during the normal build.  If that passes, the debug
build should pass without errors, so the fact that the line numbers
are wrong doesn't matter (the line numbers in stack traces *are*
correct, since they're handled by the "numberizer").